refactor(compose): move the tool-registry rebuild out of the inventory
server.go is the INVENTORY — what this process serves — and it had reached
the length where a reader stops holding it at once. It sat exactly on the
500-line ceiling, so the next handler set to be embedded was always going
to be the one that broke the gate; this change happens to be that one.

Moving the one assembly step out gives the inventory room to grow again,
and puts the registry rebuild where its own rationale can be read without
scrolling past forty embeds to reach it.
